+++ examples/noise_i2c.rs
@@ -1,0 +1,94 @@
+//! Send random raw data to the display, emulating an old untuned TV. This example `release()`s the
+//! underlying display properties struct and allows calling of the low-level `draw()` method,
+//! sending a 1024 byte buffer straight to the display.
+//!
+//! This example is for the STM32F103 "Blue Pill" board using I2C1.
+//!
+//! Wiring connections are as follows for a CRIUS-branded display:
+//!
+//! ```
+//!      Display -> Blue Pill
+//! (black)  GND -> GND
+//! (red)    +5V -> VCC
+//! (yellow) SDA -> PB9
+//! (green)  SCL -> PB8
+//! ```
+//!
+//! Run on a Blue Pill with `cargo run --example noise_i2c`. Best results when using `--release`.
+
+#![no_std]
+#![no_main]
+
+extern crate cortex_m;
+extern crate cortex_m_rt as rt;
+extern crate panic_semihosting;
+extern crate stm32f1xx_hal as hal;
+
+use cortex_m_rt::{entry, exception, ExceptionFrame};
+use embedded_graphics::{
+    pixelcolor::BinaryColor,
+    prelude::*,
+    primitives::{Circle, Rectangle, Triangle},
+    style::PrimitiveStyleBuilder,
+};
+use hal::{
+    i2c::{BlockingI2c, DutyCycle, Mode},
+    prelude::*,
+    stm32,
+};
+use rand::prelude::*;
+use ssd1306::{mode::displaymode::DisplayModeTrait, prelude::*, Builder};
+
+#[entry]
+fn main() -> ! {
+    let dp = stm32::Peripherals::take().unwrap();
+
+    let mut flash = dp.FLASH.constrain();
+    let mut rcc = dp.RCC.constrain();
+
+    let clocks = rcc.cfgr.freeze(&mut flash.acr);
+
+    let mut afio = dp.AFIO.constrain(&mut rcc.apb2);
+
+    let mut gpiob = dp.GPIOB.split(&mut rcc.apb2);
+
+    let scl = gpiob.pb8.into_alternate_open_drain(&mut gpiob.crh);
+    let sda = gpiob.pb9.into_alternate_open_drain(&mut gpiob.crh);
+
+    let i2c = BlockingI2c::i2c1(
+        dp.I2C1,
+        (scl, sda),
+        &mut afio.mapr,
+        Mode::Fast {
+            frequency: 400_000.hz(),
+            duty_cycle: DutyCycle::Ratio2to1,
+        },
+        clocks,
+        &mut rcc.apb1,
+        1000,
+        10,
+        1000,
+        1000,
+    );
+
+    let mut disp: GraphicsMode<_> = Builder::new().connect_i2c(i2c).into();
+
+    disp.init().unwrap();
+
+    let mut props = disp.release();
+
+    let mut buf = [0x00u8; 1024];
+
+    let mut rng = SmallRng::seed_from_u64(0xdead_beef_cafe_d00d);
+
+    loop {
+        rng.fill_bytes(&mut buf);
+
+        props.draw(&buf);
+    }
+}
+
+#[exception]
+fn HardFault(ef: &ExceptionFrame) -> ! {
+    panic!("{:#?}", ef);
+}